deployer/pyproject.toml
[tool.poetry]
name = "deployer"
version = "0.1.0"
license = "MPL-2.0"
description = "Upload the files that Yari built"
authors = ["MDN Devs <mdn-dev@mozilla.com>"]
repository = "https://github.com/mdn/yari"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
boto3 = "^1.13.23"
python-decouple = "^3.3"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.2"

[tool.poetry.scripts]
deployer = "deployer.main:cli"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

